+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+- add `getBucketLocation` method
+- add `getBucketCORS`, `putBucketCORS`, `deleteBucketCORS` CORS methods
+
 ## [2.1.28] - 2023-05-30
 
 ### Fixed